added methods and classes for step3&4

+++ src/edu/upenn/cit594/processor/DataProcessor.java
@@ -18,7 +18,7 @@ public class DataProcessor {
 		return filteredProperties;
 	}
 	
-	public int average (ArrayList<Property> properties, AverageCalculator averageCalculator) {
+	public Double average (ArrayList<Property> properties, AverageCalculator averageCalculator) {
 		return averageCalculator.calculateAverage(properties);
 	}
